% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_connection.R
\name{disconnect_go}
\alias{disconnect_go}
\title{Close the cached GO semsql connection}
\usage{
disconnect_go()
}
\value{
\code{NULL} invisibly.
}
\description{
Disconnects the DuckDB instance and clears the package cache.  The next
call to any query function will trigger reconnection automatically.
}
\examples{
make_go_con()
disconnect_go()
go_connection_active()

}
\seealso{
\code{\link{make_go_con}}, \code{\link{go_connection_active}}
}
